requirements.txt
requests
python-memcached
flask
boto3
